backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
anthropic
ortools
openpyxl
reportlab
python-multipart
httpx
